tests: move/drop some tests
