+++ tests/unit/fs/test_s3.py
@@ -1,7 +1,10 @@
+import importlib
+import sys
+import textwrap
+
 import pytest
 
 from dvc.config import ConfigError
-from dvc.exceptions import DvcException
 from dvc.fs.s3 import S3FileSystem
 
 bucket_name = "bucket-name"
@@ -39,13 +42,14 @@ def test_grants(dvc):
     }
     fs = S3FileSystem(dvc, config)
 
+    extra_args = fs.login_info["s3_additional_kwargs"]
     assert (
-        fs.extra_args["GrantRead"]
+        extra_args["GrantRead"]
         == "id=read-permission-id,id=other-read-permission-id"
     )
-    assert fs.extra_args["GrantReadACP"] == "id=read-acp-permission-id"
-    assert fs.extra_args["GrantWriteACP"] == "id=write-acp-permission-id"
-    assert fs.extra_args["GrantFullControl"] == "id=full-control-permission-id"
+    assert extra_args["GrantReadACP"] == "id=read-acp-permission-id"
+    assert extra_args["GrantWriteACP"] == "id=write-acp-permission-id"
+    assert extra_args["GrantFullControl"] == "id=full-control-permission-id"
 
 
 def test_grants_mutually_exclusive_acl_error(dvc, grants):
@@ -58,7 +62,7 @@ def test_grants_mutually_exclusive_acl_error(dvc, grants):
 
 def test_sse_kms_key_id(dvc):
     fs = S3FileSystem(dvc, {"url": url, "sse_kms_key_id": "key"})
-    assert fs.extra_args["SSEKMSKeyId"] == "key"
+    assert fs.login_info["s3_additional_kwargs"]["SSEKMSKeyId"] == "key"
 
 
 def test_key_id_and_secret(dvc):
@@ -71,44 +75,70 @@ def test_key_id_and_secret(dvc):
             "session_token": session_token,
         },
     )
-    assert fs.access_key_id == key_id
-    assert fs.secret_access_key == key_secret
-    assert fs.session_token == session_token
+    assert fs.login_info["key"] == key_id
+    assert fs.login_info["secret"] == key_secret
+    assert fs.login_info["token"] == session_token
 
 
-def test_get_s3_no_credentials(mocker):
-    from botocore.exceptions import NoCredentialsError
-
-    fs = S3FileSystem(None, {})
-    with pytest.raises(DvcException, match="Unable to find AWS credentials"):
-        with fs._get_s3():
-            raise NoCredentialsError
+KB = 1024
+MB = KB ** 2
+GB = KB ** 3
 
 
-def test_get_s3_connection_error(mocker):
-    from botocore.exceptions import EndpointConnectionError
+def test_s3_aws_config(tmp_dir, dvc, s3, monkeypatch):
+    config_directory = tmp_dir / ".aws"
+    config_directory.mkdir()
+    (config_directory / "config").write_text(
+        textwrap.dedent(
+            """\
+    [default]
+    s3 =
+      multipart_chunksize = 64MB
+      use_accelerate_endpoint = true
+      addressing_style = path
+    """
+        )
+    )
 
-    fs = S3FileSystem(None, {})
+    if sys.platform == "win32":
+        var = "USERPROFILE"
+    else:
+        var = "HOME"
+    monkeypatch.setenv(var, str(tmp_dir))
 
-    msg = "Unable to connect to 'AWS S3'."
-    with pytest.raises(DvcException, match=msg):
-        with fs._get_s3():
-            raise EndpointConnectionError(endpoint_url="url")
+    # Fresh import to see the effects of changing HOME variable
+    s3_mod = importlib.reload(sys.modules[S3FileSystem.__module__])
+    fs = s3_mod.S3FileSystem(dvc, s3.config)
 
-
-def test_get_s3_connection_error_endpoint(mocker):
-    from botocore.exceptions import EndpointConnectionError
-
-    fs = S3FileSystem(None, {"endpointurl": "https://example.com"})
-
-    msg = "Unable to connect to 'https://example.com'."
-    with pytest.raises(DvcException, match=msg):
-        with fs._get_s3():
-            raise EndpointConnectionError(endpoint_url="url")
+    s3_config = fs.login_info["config_kwargs"]["s3"]
+    assert s3_config["use_accelerate_endpoint"]
+    assert s3_config["addressing_style"] == "path"
+    assert fs._open_args["block_size"] == 64 * MB
 
 
-def test_get_bucket():
-    fs = S3FileSystem(None, {"url": "s3://mybucket/path"})
-    with pytest.raises(DvcException, match="Bucket 'mybucket' does not exist"):
-        with fs._get_bucket("mybucket") as bucket:
-            raise bucket.meta.client.exceptions.NoSuchBucket({}, None)
+def test_s3_aws_config_different_profile(tmp_dir, dvc, s3, monkeypatch):
+    config_file = tmp_dir / "aws_config.ini"
+    config_file.write_text(
+        textwrap.dedent(
+            """\
+    [default]
+    extra = keys
+    s3 =
+      addressing_style = auto
+      use_accelerate_endpoint = true
+      multipart_threshold = ThisIsNotGoingToBeCasted!
+    [profile dev]
+    some_extra = keys
+    s3 =
+      addressing_style = virtual
+      multipart_threshold = 2GiB
+    """
+        )
+    )
+    monkeypatch.setenv("AWS_CONFIG_FILE", config_file)
+
+    fs = S3FileSystem(dvc, {**s3.config, "profile": "dev"})
+    s3_config = fs.login_info["config_kwargs"]["s3"]
+    assert "use_accelerate_endpoint" not in s3_config
+    assert s3_config["addressing_style"] == "virtual"
+    assert fs._open_args["block_size"] == 2 * GB
